Fixed runtime bug incorrectly displaying graphics in SFML window. Added all opcode implementations except for input related opcodes. Emulator can now fully compile, load ROMS, and simulate ROM instructions, however there is a slight graphical issue that may or may not be intentional.
Version of emulator is now 0.1

+++ CPU.h
@@ -59,7 +59,7 @@ private:
   uint8_t delay_timer_;
   uint8_t sound_timer_;
   uint8_t drawflag_;
-  uint8_t index_register_;
+  uint16_t index_register_;
   uint64_t program_counter_;
   uint16_t stack_ptr_;
   uint16_t opcode_;
